+++ src/Response/Model/LayoutContent.php
@@ -17,6 +17,7 @@ use InstagramAPI\AutoPropertyMapper;
  * @method string getRelatedStyle()
  * @method TabsInfo getTabsInfo()
  * @method ThreeByFourItem getThreeByFourItem()
+ * @method TwoByTwoItem getTwoByTwoAdItem()
  * @method TwoByTwoItem getTwoByTwoItem()
  * @method bool isExploreItemInfo()
  * @method bool isFeedType()
@@ -28,6 +29,7 @@ use InstagramAPI\AutoPropertyMapper;
  * @method bool isRelatedStyle()
  * @method bool isTabsInfo()
  * @method bool isThreeByFourItem()
+ * @method bool isTwoByTwoAdItem()
  * @method bool isTwoByTwoItem()
  * @method $this setExploreItemInfo(ExploreItemInfo $value)
  * @method $this setFeedType(string $value)
@@ -39,6 +41,7 @@ use InstagramAPI\AutoPropertyMapper;
  * @method $this setRelatedStyle(string $value)
  * @method $this setTabsInfo(TabsInfo $value)
  * @method $this setThreeByFourItem(ThreeByFourItem $value)
+ * @method $this setTwoByTwoAdItem(TwoByTwoItem $value)
  * @method $this setTwoByTwoItem(TwoByTwoItem $value)
  * @method $this unsetExploreItemInfo()
  * @method $this unsetFeedType()
@@ -50,6 +53,7 @@ use InstagramAPI\AutoPropertyMapper;
  * @method $this unsetRelatedStyle()
  * @method $this unsetTabsInfo()
  * @method $this unsetThreeByFourItem()
+ * @method $this unsetTwoByTwoAdItem()
  * @method $this unsetTwoByTwoItem()
  */
 class LayoutContent extends AutoPropertyMapper
@@ -66,5 +70,6 @@ class LayoutContent extends AutoPropertyMapper
         'one_by_two_item'    => 'OneByTwoItem',
         'two_by_two_item'    => 'TwoByTwoItem',
         'three_by_four_item' => 'ThreeByFourItem',
+        'two_by_two_ad_item' => 'TwoByTwoItem',
     ];
 }