+++ src/Response/Model/TwoByTwoItem.php
@@ -10,26 +10,31 @@ use InstagramAPI\AutoPropertyMapper;
  * @method Channel getChannel()
  * @method IGTV getIgtv()
  * @method Item getMedia()
+ * @method Item getMediaOrAd()
  * @method Shopping getShopping()
  * @method bool isChannel()
  * @method bool isIgtv()
  * @method bool isMedia()
+ * @method bool isMediaOrAd()
  * @method bool isShopping()
  * @method $this setChannel(Channel $value)
  * @method $this setIgtv(IGTV $value)
  * @method $this setMedia(Item $value)
+ * @method $this setMediaOrAd(Item $value)
  * @method $this setShopping(Shopping $value)
  * @method $this unsetChannel()
  * @method $this unsetIgtv()
  * @method $this unsetMedia()
+ * @method $this unsetMediaOrAd()
  * @method $this unsetShopping()
  */
 class TwoByTwoItem extends AutoPropertyMapper
 {
     const JSON_PROPERTY_MAP = [
-        'shopping'  => 'Shopping',
-        'igtv'      => 'IGTV',
-        'channel'   => 'Channel',
-        'media'     => 'Item',
+        'shopping'        => 'Shopping',
+        'igtv'            => 'IGTV',
+        'channel'         => 'Channel',
+        'media'           => 'Item',
+        'media_or_ad'     => 'Item',
     ];
 }