+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Stable release v20.22.7
+## Date: 26/10/2022
+
+### Updates and fixes
+
+**Internal:** Added story emoji reaction sticker
+**Event:** Update `prepareAndSendExploreImpression()`
+
 # Stable release v20.22.6
 ## Date: 24/10/2022